refactor: change internal package to pkg
