+++ integrations/kgin/v1/gin-v1.go
@@ -12,8 +12,8 @@ import (
 	"sync"
 
 	"github.com/gin-gonic/gin"
-	internal "github.com/keploy/go-sdk/internal/keploy"
 	"github.com/keploy/go-sdk/keploy"
+	internal "github.com/keploy/go-sdk/pkg/keploy"
 	"go.keploy.io/server/pkg/models"
 	"go.uber.org/zap"
 )
