+++ integrations/kmongo/mdb.go
@@ -8,8 +8,8 @@ import (
 
 	"go.keploy.io/server/pkg/models"
 
-	internal "github.com/keploy/go-sdk/internal/keploy"
 	"github.com/keploy/go-sdk/keploy"
+	internal "github.com/keploy/go-sdk/pkg/keploy"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -18,9 +18,9 @@ import (
 
 // NewCollection creates and returns an instance of Collection that contains actual
 // pointer to mongo's collection. This is done in order to mock mongo CRUD operations, so that:
-//  - In keploy.MODE_RECORD mode, stores the encoded output(generated by mocked methods of mongo.Collection) into keploy's Context Deps array.
-//  - In keploy.MODE_TEST mode, decodes its stored encoded output which are present in the keploy's Context Deps array without calling mocked methods of mongo.Collection.
-//  - In "off" mode, returns the output generated after calling mocked method of mongo.Collection.
+//   - In keploy.MODE_RECORD mode, stores the encoded output(generated by mocked methods of mongo.Collection) into keploy's Context Deps array.
+//   - In keploy.MODE_TEST mode, decodes its stored encoded output which are present in the keploy's Context Deps array without calling mocked methods of mongo.Collection.
+//   - In "off" mode, returns the output generated after calling mocked method of mongo.Collection.
 //
 // cl parameter is pointer to mongo's collection instance created by (*mongo.Database).Collection
 // method. It should not be nil, else warning will logged and nil is returned.
