+++ integrations/kmongo/singleResult.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"fmt"
 
-	internal "github.com/keploy/go-sdk/internal/keploy"
 	"github.com/keploy/go-sdk/keploy"
+	internal "github.com/keploy/go-sdk/pkg/keploy"
 	"go.keploy.io/server/pkg/models"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -14,9 +14,9 @@ import (
 )
 
 // SingleResult countains instance of mongo.SingleResult to mock its methods so that:
-//  - In keploy.MODE_RECORD mode, stores the encoded output(generated by mocked methods of mongo.SingleResult) into keploy's Context Deps array.
-//  - In keploy.MODE_TEST mode, decodes its stored encoded output which are present in the keploy's Context Deps array without calling mocked methods of mongo.SingleResult.
-//  - In "off" mode, returns the output generated after calling mocked method of mongo.SingleResult.
+//   - In keploy.MODE_RECORD mode, stores the encoded output(generated by mocked methods of mongo.SingleResult) into keploy's Context Deps array.
+//   - In keploy.MODE_TEST mode, decodes its stored encoded output which are present in the keploy's Context Deps array without calling mocked methods of mongo.SingleResult.
+//   - In "off" mode, returns the output generated after calling mocked method of mongo.SingleResult.
 type SingleResult struct {
 	mongo.SingleResult
 	filter interface{}
