+++ integrations/kredis/get.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 
 	"github.com/go-redis/redis/v8"
-	internal "github.com/keploy/go-sdk/internal/keploy"
 	"github.com/keploy/go-sdk/keploy"
+	internal "github.com/keploy/go-sdk/pkg/keploy"
 
 	"go.keploy.io/server/pkg/models"
 )
