+++ integrations/ksql/v1/execerContext.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 
 	"github.com/keploy/go-sdk/integrations/ksql/ksqlErr"
-	internal "github.com/keploy/go-sdk/internal/keploy"
 	"github.com/keploy/go-sdk/keploy"
+	internal "github.com/keploy/go-sdk/pkg/keploy"
 	"go.keploy.io/server/pkg/models"
 )
