+++ integrations/ksql/v2/connBeginTx.go
@@ -8,8 +8,8 @@ import (
 
 	// v2 "github.com/keploy/go-sdk/integrations/ksql/v2"
 	"github.com/keploy/go-sdk/integrations/ksql/ksqlErr"
-	internal "github.com/keploy/go-sdk/internal/keploy"
 	"github.com/keploy/go-sdk/keploy"
+	internal "github.com/keploy/go-sdk/pkg/keploy"
 	"go.keploy.io/server/pkg/models"
 	"go.uber.org/zap"
 )
